refactor(mail): remove old mail detail and folder views, introduce mailbox display utilities
- Deleted the mail detail view and folder index view components to streamline the mail interface.
- Added a new utility module for mailbox display, providing functions for building a nested mailbox tree and handling URL segments.
- Improved handling of mailbox paths and breadcrumbs for better navigation.
