+++ app/layouts/default.vue
@@ -8,8 +8,46 @@ import DelivrIcon from '~/components/img/DelivrIcon.vue';
 import DelivrLogo from '~/components/img/DelivrLogo.vue';
 import { useSelectedMailAccountStore } from '~/composables/stores/useSelectedMailAccountStore';
 import { useUserInfoStore } from '~/composables/stores/useUserStore';
+import { buildMailboxTree, folderUrl, type MailboxTreeNode } from '~/utils/mailboxDisplay';
 
 const route = useRoute();
+
+// Original folder icon logic: match the full lowercased mailbox path.
+function folderIconFor(node: MailboxTreeNode): string {
+    const lowerPath = node.mailbox?.path.toLowerCase();
+    if (lowerPath === 'sent' || lowerPath === 'sent mail' || lowerPath === 'sent messages') return 'i-lucide-send';
+    if (lowerPath === 'drafts') return 'i-lucide-file-edit';
+    if (lowerPath === 'trash' || lowerPath === 'deleted' || lowerPath === 'deleted messages') return 'i-lucide-trash-2';
+    if (lowerPath === 'spam' || lowerPath === 'junk') return 'i-lucide-shield-alert';
+    if (lowerPath === 'archive') return 'i-lucide-archive';
+    return 'i-lucide-folder';
+}
+
+// Does the active route point at this node or any of its descendants?
+// Used to auto-expand the ancestors of the folder currently being viewed.
+function subtreeContainsActive(node: MailboxTreeNode, accountId: number): boolean {
+    if (node.mailbox && route.path === folderUrl(accountId, node.mailbox)) return true;
+    return node.children.some((child) => subtreeContainsActive(child, accountId));
+}
+
+// Map a folder-tree node to a (possibly nested) navigation menu item.
+function toNavItem(node: MailboxTreeNode, accountId: number): NavigationMenuItem {
+    const unseen = node.mailbox ? node.mailbox.status.unseen : node.unseenTotal;
+    const item: NavigationMenuItem = {
+        label: node.name,
+        icon: folderIconFor(node),
+        badge: unseen > 0 ? unseen : undefined,
+        exact: false,
+    };
+    if (node.mailbox) {
+        item.to = folderUrl(accountId, node.mailbox);
+    }
+    if (node.children.length > 0) {
+        item.children = node.children.map((child) => toNavItem(child, accountId));
+        item.defaultOpen = subtreeContainsActive(node, accountId);
+    }
+    return item;
+}
 
 const userInfoStore = useUserInfoStore();
 const user = await userInfoStore.use();
@@ -54,24 +92,13 @@ const sidebarItems = computed(() => {
         });
     }
 
-    for (const mailbox of mailboxes.value) {
-        if (mailbox.path.toLowerCase() === 'inbox') continue;
-
-        const lowerPath = mailbox.path.toLowerCase();
-        let icon = 'i-lucide-folder';
-        if (lowerPath === 'sent' || lowerPath === 'sent mail' || lowerPath === 'sent messages') icon = 'i-lucide-send';
-        else if (lowerPath === 'drafts') icon = 'i-lucide-file-edit';
-        else if (lowerPath === 'trash' || lowerPath === 'deleted' || lowerPath === 'deleted messages') icon = 'i-lucide-trash-2';
-        else if (lowerPath === 'spam' || lowerPath === 'junk') icon = 'i-lucide-shield-alert';
-        else if (lowerPath === 'archive') icon = 'i-lucide-archive';
-
-        mailItems.push({
-            label: mailbox.name,
-            icon,
-            to: currentMailAccount.value ? `/mail/${currentMailAccount.value.id}/folder/${encodeURIComponent(mailbox.path)}` : undefined,
-            badge: mailbox.status.unseen > 0 ? mailbox.status.unseen : undefined,
-            exact: false,
-        });
+    // Build the remaining folders as a nested tree (Inbox is pinned above).
+    // Hierarchy, leaf names and delimiters are all derived in buildMailboxTree.
+    const accountId = currentMailAccount.value?.id;
+    if (accountId !== undefined) {
+        for (const node of buildMailboxTree(mailboxes.value)) {
+            mailItems.push(toNavItem(node, accountId));
+        }
     }
 
     const adminItems: NavigationMenuItem[] = [
@@ -113,6 +140,10 @@ const sidebarItems = computed(() => {
             label: "API Keys",
             to: "/settings/apikeys",
             icon: "i-lucide-key",
+            // The detail/create pages (/settings/apikeys/new, /settings/apikeys/:id)
+            // are sibling routes, not children, so router-based active matching
+            // misses them. Force active on any apikeys sub-route.
+            active: route.path.startsWith("/settings/apikeys"),
         }
     ];
